* move simpleProcessMigrationTest to TaskMigrationTest * move shouldMigrateMultiLevel to BatchConfigurationTest + add pagination log checks
