+++ qa/src/test/java/io/camunda/migrator/qa/BatchConfigurationTest.java
@@ -8,6 +8,8 @@
 package io.camunda.migrator.qa;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -70,5 +72,72 @@ class BatchConfigurationTest extends RuntimeMigrationAbstractTest {
     assertThat(output.getOut()).contains("No more migrator jobs available.");
   }
 
+  @Test
+  public void shouldPaginateMultiLevelProcessModel(CapturedOutput output) {
+    // deploy processes
+    deployModels();
+
+    // given
+    runtimeService.startProcessInstanceByKey("root");
+
+    // when running runtime migration
+    runtimeMigrator.migrate();
+
+    // then
+    List<ProcessInstance> processInstances = camundaClient.newProcessInstanceSearchRequest().send().join().items();
+    assertEquals(3, processInstances.size());
+
+    Matcher matcher = Pattern.compile("Migrator jobs found: 1").matcher(output.getOut());
+    assertThat(matcher.results().count()).isEqualTo(3);
+    assertThat(output.getOut()).contains("Method: #fetchProcessInstancesToMigrate, max count: 1, offset: 0, batch size: 500");
+    assertThat(output.getOut()).contains("Method: #validateProcessInstanceState, max count: 3, offset: 0, batch size: 500");
+  }
+
+  private void deployModels() {
+    String rootId = "root";
+    String level1Id = "level1";
+    String level2Id = "level2";
+    // C7
+    var c7rootModel = org.camunda.bpm.model.bpmn.Bpmn.createExecutableProcess(rootId)
+        .startEvent("start_1")
+        .callActivity("ca_level_1")
+        .camundaIn(level1Id, level2Id)
+          .calledElement(level1Id)
+        .endEvent("end_1").done();
+    var c7level1Model = org.camunda.bpm.model.bpmn.Bpmn.createExecutableProcess(level1Id)
+        .startEvent("start_2")
+        .callActivity("ca_level_2")
+          .calledElement(level2Id)
+        .endEvent("end_2").done();
+    var c7level2Model = org.camunda.bpm.model.bpmn.Bpmn.createExecutableProcess(level2Id)
+        .startEvent("start_3")
+        .userTask("userTask_1")
+        .endEvent("end_3").done();
+
+    // C8
+    var c8rootModel = io.camunda.zeebe.model.bpmn.Bpmn.createExecutableProcess(rootId)
+        .startEvent("start_1")
+        .zeebeEndExecutionListener("migrator")
+        .callActivity("ca_level_1", c -> c.zeebeProcessId(level1Id))
+        .endEvent("end_1").done();
+    var c8level1Model = io.camunda.zeebe.model.bpmn.Bpmn.createExecutableProcess(level1Id)
+        .startEvent("start_2")
+        .zeebeEndExecutionListener("migrator")
+        .callActivity("ca_level_2", c -> c.zeebeProcessId(level2Id))
+        .endEvent("end_2").done();
+    var c8level2Model = io.camunda.zeebe.model.bpmn.Bpmn.createExecutableProcess(level2Id)
+        .startEvent("start_3")
+        .zeebeEndExecutionListener("migrator")
+        .userTask("userTask_1")
+        .endEvent("end_3").done();
+
+    repositoryService.createDeployment().addModelInstance(rootId+".bpmn", c7rootModel).deploy();
+    repositoryService.createDeployment().addModelInstance(level1Id+".bpmn", c7level1Model).deploy();
+    repositoryService.createDeployment().addModelInstance(level2Id+".bpmn", c7level2Model).deploy();
+
+    camundaClient.newDeployResourceCommand().addProcessModel(c8rootModel, rootId+".bpmn").send().join();
+    camundaClient.newDeployResourceCommand().addProcessModel(c8level1Model, level1Id+".bpmn").send().join();
+    camundaClient.newDeployResourceCommand().addProcessModel(c8level2Model, level2Id+".bpmn").send().join();
+  }
 
 }
